Move some of the chunk management logic from TerrainController to a new ChunkController class
