+++ Assets/Scripts/Terrain/ChunkController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 51ec75454a4f3bb4b88c353af0492301
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
